evaluation/requirements.txt
git+https://github.com/explodinggradients/ragas@v0.1.9
deepeval~=0.21.57
langchain-openai~=0.1.8
pytest
pytest-asyncio
aiolimiter
chromadb-client~=0.4.25.dev0

